Added passenger service class and made changes to0 passengerspring application
